feat: support system-metrics v3 / queue-metrics v3 chain
Bump cboxdk/laravel-queue-metrics to ^3.0 and widen CPU core fields
from int to float across ClusterManagerState, AutoscaleManager, and
CapacityCalculator. This enables fractional core counts from cgroup
limits (e.g. 0.5 cores in containers) to flow through heartbeats,
cluster summaries, and capacity calculations without truncation.

+++ src/Manager/AutoscaleManager.php
@@ -269,7 +269,7 @@ final class AutoscaleManager
             availableWorkerCapacity: max($capacity->finalMaxWorkers - $this->pool->totalCount(), 0),
             capacityLimiter: $this->clusterString($capacity->limitingFactor, 'unknown'),
             cpuPercent: $this->clusterFloat($cpuDetails['current_cpu_percent'] ?? 0.0),
-            cpuCores: is_numeric($cpuDetails['total_cores'] ?? null) ? (int) $cpuDetails['total_cores'] : 0,
+            cpuCores: is_numeric($cpuDetails['total_cores'] ?? null) ? (float) $cpuDetails['total_cores'] : 0.0,
             cpuUsableCores: is_numeric($cpuDetails['usable_cores'] ?? null) ? (int) $cpuDetails['usable_cores'] : 0,
             cpuReservedCores: is_numeric($cpuDetails['reserve_cores'] ?? null) ? (int) $cpuDetails['reserve_cores'] : 0,
             memoryPercent: $this->clusterFloat($memoryDetails['current_memory_percent'] ?? 0.0),
