+++ src/Scaling/Calculators/CapacityCalculator.php
@@ -21,7 +21,7 @@ class CapacityCalculator
 
     private ?float $cachedTotalMemoryMb = null;
 
-    private ?int $cachedAvailableCores = null;
+    private ?float $cachedAvailableCores = null;
 
     private ?float $cacheTimestamp = null;
